feat: PWA installabile, refactor service layer e hardening sicurezza (v0.2.0)
PWA:
- manifest.json + service worker (sw.js) presenti su tutte le pagine d'ingresso
  (login, splash, dashboard); l'app è ora installabile su Android/iOS/desktop
- splash screen con health-check del backend

Refactor:
- services/docker.py -> package modulare services/docker/ suddiviso per dominio
  (containers, images_updates, networks, volumes, system, events, base)

Sicurezza:
- modalità sicura coerente: delete/kill container richiedono conferma (403 senza confirm),
  con dialog di conferma in UI (singolo e bulk)
- rate limiting login/API basato sull'IP reale dietro proxy (CF-Connecting-IP / XFF)
- protezione CSRF, timeout di sessione, redazione dei valori sensibili nei log
- chiuso edge case open-redirect (/\host) nel parametro next

Fix:
- le azioni sui container non riportano più un falso "successo" quando Docker fallisce
- remove_image propaga gli errori con feedback chiaro in UI
- corretto template splash (Jinja {{ }}) e validazione input frequenza aggiornamenti
- log di debug del full-update convertiti da print() a logger

Docs:
- README aggiornato (PWA, architettura, sicurezza, struttura progetto, endpoint)
- aggiunto CHANGELOG.md
