+++ d2ha/static/manifest.json
@@ -1,0 +1,33 @@
+{
+  "id": "/",
+  "name": "Docker 2 Home Assistant",
+  "short_name": "D2HA",
+  "description": "Dashboard to manage Docker containers for Home Assistant",
+  "start_url": "/",
+  "scope": "/",
+  "display": "standalone",
+  "lang": "it",
+  "dir": "ltr",
+  "background_color": "#12141a",
+  "theme_color": "#1a1d24",
+  "icons": [
+    {
+      "src": "/static/img/icon-192.png",
+      "sizes": "192x192",
+      "type": "image/png",
+      "purpose": "any"
+    },
+    {
+      "src": "/static/img/icon-512.png",
+      "sizes": "512x512",
+      "type": "image/png",
+      "purpose": "any"
+    },
+    {
+      "src": "/static/img/icon-512.png",
+      "sizes": "512x512",
+      "type": "image/png",
+      "purpose": "maskable"
+    }
+  ]
+}
