+++ d2ha/templates/autodiscovery.html
@@ -1,26 +1,9 @@
-<!DOCTYPE html>
-<html lang="{{ get_current_lang() }}">
+{% extends "layouts/app_base.html" %}
 
-<head>
-  <meta charset="UTF-8" />
-  <meta name="viewport" content="width=device-width, initial-scale=1.0" />
-  <title>D2HA – MQTT</title>
-  <link rel="stylesheet" href="{{ url_for('static', filename='css/main.css') }}">
-  <link rel="icon" type="image/png" href="{{ url_for('static', filename='img/favicon.png') }}">
-  {% include 'partials/notifications_styles.html' %}
-</head>
+{% block title %}D2HA – MQTT{% endblock %}
 
-<body class="theme-{{ get_current_theme() }}" data-safe-mode="{{ '1' if safe_mode_enabled else '0' }}"
-  data-performance-mode="{{ '1' if performance_mode_enabled else '0' }}"
-  data-debug-mode="{{ '1' if debug_mode_enabled else '0' }}">
-  {% include 'partials/app_sidebar.html' %}
-  <div class="main-wrapper">
-    {% include 'partials/app_topbar.html' %}
-
-
-  {% include 'partials/flash_messages.html' %}
-
-  <main>
+{% block content %}
+<main>
     <section class="mqtt-status-bar">
       <div class="title">
         <h3>MQTT</h3>
@@ -179,8 +162,10 @@
       </div>
     </div>
   </div>
-  {% include 'partials/notifications_script.html' %}
-  <script>
+{% endblock %}
+
+{% block scripts_extra %}
+<script>
     const stackToggles = Array.from(document.querySelectorAll('[data-stack-toggle]'));
     const deselectButtons = Array.from(document.querySelectorAll('.deselect-stack'));
     const autodiscoveryForm = document.getElementById('autodiscovery-form');
@@ -328,7 +313,4 @@
       applySafeModeLock(!!event.detail?.enabled);
     });
   </script>
-  </div>
-</body>
-
-</html>
+{% endblock %}
