+++ d2ha/templates/layouts/auth_base.html
@@ -5,6 +5,7 @@
   <meta charset="UTF-8" />
   <meta name="viewport" content="width=device-width, initial-scale=1.0" />
   <title>{% block title %}D2HA{% endblock %}</title>
+  {% include 'partials/pwa_head.html' %}
   <link rel="stylesheet" href="{{ url_for('static', filename='css/auth.css') }}">
   <link rel="icon" type="image/png" href="{{ url_for('static', filename='img/favicon.png') }}">
 </head>
@@ -115,6 +116,7 @@
     initCompactToggle();
   </script>
   {% block auth_scripts %}{% endblock %}
+  {% include 'partials/pwa_register.html' %}
 </body>
 
 </html>
