+++ d2ha/templates/partials/app_mobile_nav.html
@@ -1,0 +1,99 @@
+<nav class="mobile-bottom-nav" aria-label="Mobile Navigation">
+  <a href="{{ url_for('ui.index') }}" class="mobile-nav-item {% if request.endpoint == 'ui.index' %}active{% endif %}" title="{{ t('nav.home') }}">
+    <div class="mobile-nav-icon">
+      <svg viewBox="0 0 24 24" fill="none" stroke="currentColor" stroke-width="2" stroke-linecap="round" stroke-linejoin="round"><path d="M3 9l9-7 9 7v11a2 2 0 0 1-2 2H5a2 2 0 0 1-2-2z"></path><polyline points="9 22 9 12 15 12 15 22"></polyline></svg>
+    </div>
+    <span class="mobile-nav-label">{{ t('nav.home') }}</span>
+  </a>
+
+  <a href="{{ url_for('ui.containers_view') }}" class="mobile-nav-item {% if request.endpoint == 'ui.containers_view' %}active{% endif %}" title="{{ t('nav.containers') }}">
+    <div class="mobile-nav-icon">
+      <svg viewBox="0 0 24 24" fill="none" stroke="currentColor" stroke-width="2" stroke-linecap="round" stroke-linejoin="round"><rect x="2" y="7" width="20" height="14" rx="2" ry="2"></rect><path d="M16 21V5a2 2 0 0 0-2-2h-4a2 2 0 0 0-2 2v16"></path></svg>
+    </div>
+    <span class="mobile-nav-label">{{ t('nav.containers') }}</span>
+  </a>
+
+  <a href="{{ url_for('ui.images_view') }}" class="mobile-nav-item {% if request.endpoint == 'ui.images_view' %}active{% endif %}" title="{{ t('nav.images') }}">
+    <div class="mobile-nav-icon">
+      <svg viewBox="0 0 24 24" fill="none" stroke="currentColor" stroke-width="2" stroke-linecap="round" stroke-linejoin="round"><rect x="3" y="3" width="18" height="18" rx="2" ry="2"></rect><circle cx="8.5" cy="8.5" r="1.5"></circle><polyline points="21 15 16 10 5 21"></polyline></svg>
+    </div>
+    <span class="mobile-nav-label">{{ t('nav.images') }}</span>
+  </a>
+
+  <button type="button" class="mobile-nav-item mobile-more-btn" aria-expanded="false" aria-controls="mobileMoreMenu" aria-label="Altro">
+    <div class="mobile-nav-icon">
+      <svg viewBox="0 0 24 24" fill="none" stroke="currentColor" stroke-width="2" stroke-linecap="round" stroke-linejoin="round"><line x1="3" y1="12" x2="21" y2="12"></line><line x1="3" y1="6" x2="21" y2="6"></line><line x1="3" y1="18" x2="21" y2="18"></line></svg>
+    </div>
+    <span class="mobile-nav-label">Altro</span>
+  </button>
+</nav>
+
+<div class="mobile-more-menu" id="mobileMoreMenu" aria-hidden="true">
+  <div class="mobile-more-header">
+    <h2>Altro</h2>
+    <button type="button" class="mobile-more-close" aria-label="Chiudi">&times;</button>
+  </div>
+  <div class="mobile-more-content">
+    <a href="{{ url_for('ui.volumes_view') }}" class="mobile-more-link {% if request.endpoint == 'ui.volumes_view' %}active{% endif %}">
+      <svg viewBox="0 0 24 24" fill="none" stroke="currentColor" stroke-width="2" stroke-linecap="round" stroke-linejoin="round"><ellipse cx="12" cy="5" rx="9" ry="3"></ellipse><path d="M21 12c0 1.66-4 3-9 3s-9-1.34-9-3"></path><path d="M3 5v14c0 1.66 4 3 9 3s9-1.34 9-3V5"></path></svg>
+      {{ t('nav.volumes') }}
+    </a>
+    <a href="{{ url_for('ui.networks_view') }}" class="mobile-more-link {% if request.endpoint == 'ui.networks_view' %}active{% endif %}">
+      <svg viewBox="0 0 24 24" fill="none" stroke="currentColor" stroke-width="2" stroke-linecap="round" stroke-linejoin="round"><circle cx="12" cy="12" r="3"></circle><line x1="12" y1="15" x2="12" y2="21"></line><line x1="12" y1="9" x2="12" y2="3"></line><line x1="15" y1="12" x2="21" y2="12"></line><line x1="9" y1="12" x2="3" y2="12"></line></svg>
+      {{ t('nav.networks') }}
+    </a>
+    <a href="{{ url_for('ui.updates') }}" class="mobile-more-link {% if request.endpoint == 'ui.updates' %}active{% endif %}">
+      <svg viewBox="0 0 24 24" fill="none" stroke="currentColor" stroke-width="2" stroke-linecap="round" stroke-linejoin="round"><polyline points="23 4 23 10 17 10"></polyline><polyline points="1 20 1 14 7 14"></polyline><path d="M3.51 9a9 9 0 0 1 14.85-3.36L23 10M1 14l4.64 4.36A9 9 0 0 0 20.49 15"></path></svg>
+      {{ t('nav.updates') }}
+      {% if global_update_count and global_update_count > 0 %}
+      <span class="badge" style="margin-left: auto;">{{ global_update_count }}</span>
+      {% endif %}
+    </a>
+    <a href="{{ url_for('ui.events_view') }}" class="mobile-more-link {% if request.endpoint == 'ui.events_view' %}active{% endif %}">
+      <svg viewBox="0 0 24 24" fill="none" stroke="currentColor" stroke-width="2" stroke-linecap="round" stroke-linejoin="round"><path d="M21.5 2v6h-6M21.34 15.57a10 10 0 1 1-.59-9.5l5.67-5.67"></path></svg>
+      {{ t('nav.events') }}
+    </a>
+    <a href="{{ url_for('ui.autodiscovery_view') }}" class="mobile-more-link {% if request.endpoint == 'ui.autodiscovery_view' %}active{% endif %}">
+      <svg viewBox="0 0 24 24" fill="none" stroke="currentColor" stroke-width="2" stroke-linecap="round" stroke-linejoin="round"><path d="M4 11a9 9 0 0 1 9 9"></path><path d="M4 4a16 16 0 0 1 16 16"></path><circle cx="5" cy="19" r="1"></circle></svg>
+      {{ t('nav.autodiscovery') }}
+    </a>
+    
+    <div class="mobile-more-divider"></div>
+    
+    <a href="{{ url_for('auth.security_settings') }}" class="mobile-more-link {% if request.endpoint == 'auth.security_settings' %}active{% endif %}">
+      <svg viewBox="0 0 24 24" fill="none" stroke="currentColor" stroke-width="2" stroke-linecap="round" stroke-linejoin="round"><circle cx="12" cy="12" r="3"></circle><path d="M19.4 15a1.65 1.65 0 0 0 .33 1.82l.06.06a2 2 0 0 1 0 2.83 2 2 0 0 1-2.83 0l-.06-.06a1.65 1.65 0 0 0-1.82-.33 1.65 1.65 0 0 0-1 1.51V21a2 2 0 0 1-2 2 2 2 0 0 1-2-2v-.09A1.65 1.65 0 0 0 9 19.4a1.65 1.65 0 0 0-1.82.33l-.06.06a2 2 0 0 1-2.83 0 2 2 0 0 1 0-2.83l.06-.06a1.65 1.65 0 0 0 .33-1.82 1.65 1.65 0 0 0-1.51-1H3a2 2 0 0 1-2-2 2 2 0 0 1 2-2h.09A1.65 1.65 0 0 0 4.6 9a1.65 1.65 0 0 0-.33-1.82l-.06-.06a2 2 0 0 1 0-2.83 2 2 0 0 1 2.83 0l.06.06a1.65 1.65 0 0 0 1.82.33H9a1.65 1.65 0 0 0 1-1.51V3a2 2 0 0 1 2-2 2 2 0 0 1 2 2v.09a1.65 1.65 0 0 0 1 1.51 1.65 1.65 0 0 0 1.82-.33l.06-.06a2 2 0 0 1 2.83 0 2 2 0 0 1 0 2.83l-.06.06a1.65 1.65 0 0 0-.33 1.82V9a1.65 1.65 0 0 0 1.51 1H21a2 2 0 0 1 2 2 2 2 0 0 1-2 2h-.09a1.65 1.65 0 0 0-1.51 1z"></path></svg>
+      Impostazioni
+    </a>
+    <a href="{{ url_for('auth.logout') }}" class="mobile-more-link text-danger">
+      <svg viewBox="0 0 24 24" fill="none" stroke="currentColor" stroke-width="2" stroke-linecap="round" stroke-linejoin="round"><path d="M9 21H5a2 2 0 0 1-2-2V5a2 2 0 0 1 2-2h4"></path><polyline points="16 17 21 12 16 7"></polyline><line x1="21" y1="12" x2="9" y2="12"></line></svg>
+      Esci
+    </a>
+  </div>
+</div>
+
+<script>
+  (function() {
+    const moreBtn = document.querySelector('.mobile-more-btn');
+    const moreMenu = document.getElementById('mobileMoreMenu');
+    const closeBtn = document.querySelector('.mobile-more-close');
+
+    if (moreBtn && moreMenu && closeBtn) {
+      function openMenu() {
+        moreMenu.classList.add('active');
+        moreMenu.setAttribute('aria-hidden', 'false');
+        moreBtn.setAttribute('aria-expanded', 'true');
+        document.body.style.overflow = 'hidden';
+      }
+
+      function closeMenu() {
+        moreMenu.classList.remove('active');
+        moreMenu.setAttribute('aria-hidden', 'true');
+        moreBtn.setAttribute('aria-expanded', 'false');
+        document.body.style.overflow = '';
+      }
+
+      moreBtn.addEventListener('click', openMenu);
+      closeBtn.addEventListener('click', closeMenu);
+    }
+  })();
+</script>
