+++ d2ha/templates/partials/app_topbar.html
@@ -11,10 +11,7 @@
 </script>
 <header class="app-topbar">
   <div class="topbar-left">
-    <button class="mobile-drawer-tab" id="mobileDrawerTab" type="button" aria-expanded="false" aria-controls="primaryNav"
-      aria-label="Apri o chiudi la barra laterale">
-      <span class="mobile-drawer-tab-icon" aria-hidden="true">☰</span>
-    </button>
+    <h1 class="mobile-topbar-title">D2HA</h1>
   </div>
   <div class="topbar-right">
     <div class="backend-area">
