+++ d2ha/templates/partials/pwa_head.html
@@ -1,0 +1,7 @@
+<meta name="theme-color" content="#1a1d24">
+<meta name="apple-mobile-web-app-capable" content="yes">
+<meta name="mobile-web-app-capable" content="yes">
+<meta name="apple-mobile-web-app-status-bar-style" content="black-translucent">
+<meta name="apple-mobile-web-app-title" content="D2HA">
+<link rel="apple-touch-icon" href="{{ url_for('static', filename='img/icon-192.png') }}">
+<link rel="manifest" href="{{ url_for('static', filename='manifest.json') }}">
